Rework saving of the PushNotificationConfig object : there was a conflict between the 'push_parameters' string that contains the full configuration, and other more recent parts such as 'voip_token' or 'remote_token'. Now, only keep 'push_parameters', and extends the parsing of the string to fully setup all the necessary fields od the PushNotificationConfig

+++ src/utils/time-utils.mm
@@ -33,7 +33,7 @@ time_t iso8601ToTimeApple(std::string iso8601DateTime) {
 }
 
 std::string timeToIso8601Apple(time_t t) {
-	NSDate* nsDateTime = [NSDate dateWithTimeIntervalSince1970:t];
+	NSDate* nsDateTime = [NSDate dateWithTimeIntervalSince1970:(double)t];
 	NSISO8601DateFormatter* dateFormatter = [[NSISO8601DateFormatter alloc] init];
 	NSString* dateString = [dateFormatter stringFromDate:nsDateTime];
 	return std::string([dateString UTF8String]);
